Remove the rtc framework
No longer any challengers to hwloc, so streamline the
binding support by re-integrating it back into the
odls framework.

(from upstream commit add324c9f3b3d9ddb8301fbcd53ae1932de41039)
