+++ src/mca/odls/base/Makefile.am
@@ -13,7 +13,7 @@
 #                         All rights reserved
 # Copyright (c) 2019      Intel, Inc.  All rights reserved.
 # Copyright (c) 2020      Cisco Systems, Inc.  All rights reserved
-# Copyright (c) 2022      Nanook Consulting.  All rights reserved.
+# Copyright (c) 2022-2024 Nanook Consulting  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -27,6 +27,7 @@ headers += \
 libprtemca_odls_la_SOURCES += \
         base/odls_base_frame.c \
         base/odls_base_select.c \
-        base/odls_base_default_fns.c
+        base/odls_base_default_fns.c \
+        base/odls_base_bind.c
 
 dist_prtedata_DATA += base/help-prte-odls-base.txt
