--- src/mca/rtc/base/Makefile.am
@@ -1,20 +1,0 @@
-#
-# Copyright (c) 2014-2019 Intel, Inc.  All rights reserved.
-# Copyright (c) 2020      Cisco Systems, Inc.  All rights reserved
-# Copyright (c) 2022      Nanook Consulting.  All rights reserved.
-# $COPYRIGHT$
-#
-# Additional copyrights may follow
-#
-# $HEADER$
-#
-
-headers += \
-        base/base.h
-
-libprtemca_rtc_la_SOURCES += \
-        base/rtc_base_frame.c \
-        base/rtc_base_select.c \
-        base/rtc_base_stubs.c
-
-dist_prtedata_DATA = base/help-prte-rtc-base.txt
